drfbackend/requirements.txt
djangorestframework==3.13.1
psycopg2-binary==2.8.6
Django==4.0.6
python-dotenv==0.20.0